+++ ipympl_interactions/__init__.py
@@ -1,0 +1,3 @@
+from .jupyter import *
+from .generic import *
+from .utils import *
