+++ setup.py
@@ -1,8 +1,8 @@
-from setuptools import setup
+from setuptools import setup, find_packages
 setup(
     name="ipympl_interactions",
     version="0.1",
-    py_modules=['ipympl_interactions'],
+    packages=find_packages(),
     install_requires = [
         'ipympl>=0.5.6',
         # implicitly picks up a ipywidgets dep by way of ipympl
